+++ database/migrations/2025_12_07_144752_remove_url_from_galleries_table.php
@@ -11,10 +11,13 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('galleries', function (Blueprint $table) {
-            $table->dropColumn('url');
-        });
+        if (Schema::hasColumn('galleries', 'url')) {
+            Schema::table('galleries', function (Blueprint $table) {
+                $table->dropColumn('url');
+            });
+        }
     }
+
 
     /**
      * Reverse the migrations.
